fix three major bugs

+++ Assets/Scripts/Beehive.cs
@@ -50,7 +50,7 @@ public class Beehive : MonoBehaviour, IComparable
         }
         
         float surplus = Mathf.Pow(0.9f, inventory.GetValueOrDefault(pollen) / (outFlow - inFlow));
-        float baseValue = recipe.currentPrice / recipe.GetRequiredPollenCounts().Values.Sum();
+        float baseValue = (float)recipe.currentPrice / recipe.GetRequiredPollenCounts().Values.Sum();
         return surplus * baseValue;
     }
